requirements.txt
streamlit
numpy
pandas
plotly
xlsxwriter
fpdf2
kaleido